Refactor String extensions and update version to 1.3.4
Refactored String extension functions to improve readability and leverage Kotlin's extension features. Changed coordinate types in Config from int to double for better precision. Updated project version to 1.3.4 in build files.

Took 8 minutes

+++ src/main/java/com/creepyx/creepybase/config/Config.java
@@ -130,9 +130,9 @@ public class Config extends CustomConfig {
 	@Override
 	public @Nullable Location getLocation(@NotNull String path) {
 		String world = this.getString(path + ".World");
-		int x = this.getInt(path + ".X");
-		int y = this.getInt(path + ".Y");
-		int z = this.getInt(path + ".Z");
+		double x = this.getDouble(path + ".X");
+		double y = this.getDouble(path + ".Y");
+		double z = this.getDouble(path + ".Z");
 		float yaw = (float) this.getDouble(path + ".Yaw");
 		float pitch = (float) this.getDouble(path + ".Pitch");
